refactor(replay): extract playerInspectorLogsLogic
playerInspectorLogic was pushing past 1500 lines. Pull the logs-specific
actions, reducers, loaders, and the logs query builder into a dedicated
playerInspectorLogsLogic that the parent connects to.

- Owns: MAX_LOG_ENTRIES, buildSessionLogsQuery, the logs loader
  (loadLogs/loadMoreLogs), and logsHasMore/logsNextCursor/logsLoadError
  /logsInitialLoadRequested state.
- Exposes a readyToLoadLogs selector so the parent's loadEventsSuccess
  listener stays a one-liner.
- playerInspectorLogic connects to the new logic for values and actions,
  so external consumers (PlayerInspectorList, PlayerInspectorControls)
  keep the same API.

Generated-By: PostHog Code
Task-Id: ecdba672-1e61-49d1-b387-20d712bbe67b
